Move data preprocessing to new file, use pickled data for app
